// src/app/generated/player-api/model/editTeamPermissionCommand.ts
/**
 * Player API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface EditTeamPermissionCommand { 
    name?: string | null;
    description?: string | null;
}